app-development/requirements.txt
mcp[cli]
pytest
pytest-asyncio
google-generativeai
GitPython==3.1.40
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
PyGithub==2.1.1
aiosqlite==0.19.0
httpx==0.27.0
aiofiles==23.2.1
